Conditional-Mean-Var-Models
Financial econometrics project analyzing the time series of Home Depot's returns, we first applied classical conditional mean and variance models, including an ARMA(3,3) for the mean and an EGARCH(1,1) for volatility. However, due to unsatisfactory in-sample and out-of-sample performance, we explored alternative approaches. A decision tree—while unconventional for modeling return volatility—was tested for its ability to capture non-linear patterns. We also implemented a multilayer perceptron (MLP) as a deep learning method. Despite these efforts, achieving robust out-of-sample performance proved to be particularly challenging, highlighting the difficulty of modeling and forecasting financial return dynamics.